% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprint-package.R
\docType{package}
\name{imprint}
\alias{imprint}
\alias{imprint-package}
\title{Create Customized 'ggplot2' and 'R Markdown' Themes for Your Organization}
\description{
Create Customized 'ggplot2' and 'R Markdown' Themes for Your Organization
}
\author{
Bob Rudis (bob@rud.is)
}
