% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{primary_pal}
\alias{primary_pal}
\title{Intended to be your org's main color palette, example first palette from SB Admin 2 bootstrap theme. https://github.com/BlackrockDigital/startbootstrap-sb-admin-2}
\usage{
primary_pal()
}
\description{
Intended to be your org's main color palette, example first palette from SB Admin 2 bootstrap theme. https://github.com/BlackrockDigital/startbootstrap-sb-admin-2
}
\examples{
library(scales)
scales::show_col(primary_pal()(9))
}
